+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Adding the `distance/metrics/bag` namespace.
 * Adding the `distance/metrics/lcs` namespace.
 * Adding the `distance/metrics/length` namespace.
+* Adding the `distance/metrics/mlipns` namespace.
 * Adding the `distance/metrics/prefix` namespace.
 * Adding the `distance/metrics/smith-waterman` namespace.
 * Adding the `distance/metrics/suffix` namespace.